popraveno logout za mywatchlist i statistics

+++ src/main/webapp/app/scripts/controllers/mywatchlistController.js
@@ -1,61 +1,67 @@
 TvShowsApp.controller('mywatchlistController', ['$scope', 'getWatchlistService', 'removeFromWatchlistService', '$cookies', '$location',
-function($scope, getWatchlistService, removeFromWatchlistService, $cookies, $location){
+                                                 '$rootScope',
+function($scope, getWatchlistService, removeFromWatchlistService, $cookies, $location, $rootScope){
 	$scope.results = [];
 	$scope.loading = true;
 	
-	var checkIfLoggedIn = function(){
-        
-        var token = $cookies.get('token');
-        
-        if (angular.isDefined(token)){
-            
-            $scope.username = "";
-            
-            for (var i=1; i<token.length; i++){
-                  
-                if (token[i] == ':')
-                    break;
-                
-                $scope.username += token[i];
-              
-            }
-            
-            return true;
-        }
-        else
-            return false;
-    };
+	if ($rootScope.loginAction == false) {
+       
+          $location.path("/login");
+       
+      }
+	else{
 	
-	if(!checkIfLoggedIn){
-		$location.path('/login');
-	}
-
-    
-	
-	
-	var loadMovies = function(){
-		getWatchlistService.query(function(data){
-			$scope.results = data;
-		});
+		var checkIfLoggedIn = function(){
+	        
+	        var token = $cookies.get('token');
+	        
+	        if (angular.isDefined(token)){
+	            
+	            $scope.username = "";
+	            
+	            for (var i=1; i<token.length; i++){
+	                  
+	                if (token[i] == ':')
+	                    break;
+	                
+	                $scope.username += token[i];
+	              
+	            }
+	            
+	            return true;
+	        }
+	        else
+	            return false;
+	    };
 		
-		$scope.loading = false;
-	};
+		if(!checkIfLoggedIn){
+			$location.path('/login');
+		}
 	
-	loadMovies();
+	    
+		
+		
+		var loadMovies = function(){
+			getWatchlistService.query(function(data){
+				$scope.results = data;
+			});
+			
+			$scope.loading = false;
+		};
+		
+		loadMovies();
+		
+		$scope.removeFromWatchlist = function(tvShow, index){
+	  
+		$scope.results.splice(index, 1);	
+	  	  removeFromWatchlistService.remove($.param({
+	            
+	            tmdb_id: tvShow.tmdbId
+	            
+	            }));
+	  	  
+	    };
 	
-	$scope.removeFromWatchlist = function(tvShow, index){
-  
-	$scope.results.splice(index, 1);	
-  	  removeFromWatchlistService.remove($.param({
-            
-            tmdb_id: tvShow.tmdbId
-            
-            }));
-  	  
-    };
-	
-	
-	
-	
-}                                                
-]);
+	}	
+
+}]);
